Kill the gui after 10s
So as of the last commit, travis seems to have no problem building and running
the particle example. But because it's a graphical program, it won't exit
until the user closes it, which will cause travis timeouts and feels like poor
manners. This script should kill it after a brief delay to get things over
with. Credit to
http://stackoverflow.com/questions/3510673/find-and-kill-a-process-in-one-line-using-bash-and-regex