Add drone configs and badge
It looks like drone.io is ignoring the .drone.yml in the repository. I'm
guessing that they use the web interface to generate one and add it on their
end. I added a screenshot of what the working configuration looks like, and I
also added the badge to the readme.